Move messageCreate to own file
